refactor: change file name sse.rs to service.rs
